requirements.txt
fastapi>=0.109,<0.110
uvicorn[standard]>=0.27,<0.28
python-dotenv>=1.0,<2.0
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0

# LangChain ecosystem (0.1 generation only)
langchain>=0.1.20,<0.2
langchain-core>=0.1.52,<0.2
langchain-community>=0.0.38,<0.1
langchain-text-splitters>=0.0.1,<0.1
langgraph>=0.0.51,<0.1
langchain-pinecone>=0.0.1,<0.1
langchain-openai>=0.0.8,<0.1
langsmith>=0.1.147,<0.2

# OpenAI
openai>=1.30,<2.0
tiktoken>=0.5,<1.0

# Pinecone
pinecone[grpc]>=3.0,<4.0

# Utils
httpx>=0.25,<0.26
aiofiles>=23.2,<24
requests>=2.31,<3
numpy>=1.24,<2
python-multipart>=0.0.6,<0.1
